Added the file not found block

+++ week1/code/ConcatenateTwoFiles.sh
@@ -4,7 +4,13 @@ then
     echo "Please input the two files to be concatenated and the name of the output file"
     exit 1
 fi
+if [ ! -f "$1" ] || [ ! -f "$2" ]
+then 
+    echo "One of the files does not exist"
+    exit 2
+fi
 cat $1 > $3
 cat $2 >> $3
 echo "The merged file is"
 cat $3
+exit 0
